requirements.txt
openpyxl>=3.1.0
notion-client>=2.0.0
python-dotenv>=1.0.0
nbformat>=5.9.0
pandas>=1.5.0
uuid>=1.0.0
openpyxl>=3.1.0
pyarrow>=6.0.0